+++ index.html
@@ -4,7 +4,11 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React</title>
+
+    <meta name="theme-color" content="#000000" />
+    
+
+    <title>ClimaMundial</title>
   </head>
   <body>
     <div id="root"></div>
