+++ src/App.css
@@ -1,1 +1,53 @@
+.App {
+  text-align: center;
+}
 
+.App-logo {
+  height: 40vmin;
+  pointer-events: none;
+}
+
+@media (prefers-reduced-motion: no-preference) {
+  .App-logo {
+    animation: App-logo-spin infinite 20s linear;
+  }
+}
+
+.App-header {
+  background-color: #282c34;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(20px + 5vmin);
+  color: white;
+}
+
+.App-link {
+  color: #61dafb;
+}
+
+@keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
+
+section {
+  color: #fff;
+  text-shadow: 4px 4px rgba(56, 70, 73, 0.5);
+  background-color: #3e494b71;
+  height: 100vh;  
+}
+.navBg {
+  background-color: #63aa92;
+}
+
+.me-auto {
+  font-size: 35px;
+  text-decoration-color:  white;
+}
