+++ src/components/Contacto.jsx
@@ -1,0 +1,8 @@
+const Contacto = () => {
+    return(
+        <div>
+            <h1>Contacto</h1>
+        </div>
+    )
+}
+export default Contacto
