+++ src/components/Inicio.jsx
@@ -1,0 +1,13 @@
+
+import { Card} from './cardApp';
+
+ 
+const Inicio = () => {
+    return(
+        <div>
+            <Card/> 
+        </div>
+    )
+}
+export default Inicio
+
