+++ src/components/Nosotros.jsx
@@ -1,0 +1,16 @@
+const Nosotros = () => {
+    return(
+        <div>
+            <h1> Nosotros</h1>
+            <h2>¿QUIENES SOMOS?​
+
+
+Somos una comunidad sin fines de lucro, de mujeres venezolanas con una cosa en común: "Pasión por la Tecnología".​
+Hoy traemos para TI, una App Web, que nos permitirà conocer el estado del Tiempo a Nivel Mundial, "Un Salto al Clima", inspirado en nuestras raìces Venezolanas, para el Mundo entero. ​
+Tenemos la intención de llevar la inclusión de todas las mujeres sin importar su raza, religión, credo, edad, etnia, Etc, para la Construcciòn de un mundo igualitario, e
+
+</h2>
+        </div>
+    )
+}
+export default Nosotros
